+++ cayenne-project/src/main/java/org/apache/cayenne/project/upgrade/handlers/v11/ToDepPkToFkUpdater.java
@@ -17,7 +17,7 @@
  *  under the License.
  ****************************************************************/
 
-package org.apache.cayenne.project.upgrade.utils;
+package org.apache.cayenne.project.upgrade.handlers.v11;
 
 import org.apache.cayenne.project.upgrade.UpgradeUnit;
 import org.w3c.dom.Element;
